requirements.txt
torch
torchvision
numpy
Pillow
tqdm
streamlit
ftfy
regex
git+https://github.com/openai/CLIP.git
